[GH-94]: dividir server/index.js en routers por dominio
- server/middleware/auth.js: authGuard y adminGuard compartidos
- server/routes/auth.js: POST /register, POST /loginuser
- server/routes/clientes.js: CRUD /cliente
- server/routes/reservas.js: CRUD /reserva, check-in, cenas, XML
- server/routes/facturas.js: POST /factura, POST /factura/preview
- server/routes/menu.js: CRUD /menu, GET /menu/pdf (público)
- server/routes/precios.js: GET/PUT /precios-base (rutas específicas antes de /:id)
- server/routes/usuarios.js: CRUD /usuarios (solo admin)
- server/routes/bookingSync.js: /booking-sync/*
- server/routes/uploadBooking.js: POST /upload-booking
- server/routes/dni.js: POST /parse-dni
- server/index.js: reducido a ~35 líneas de setup y montaje de routers
